[artf1343166]: Fix for Clang-tidy warnings
   Issue: enum 'xxxx' uses a larger base type ('uint16_t' (aka 'unsigned short'), size: 2 bytes) than necessary for its value set, consider using 'std::uint8_t' (1 byte) as the base type to reduce its size

+++ nfc_reader/aidl/2.0/NfcExtns.cpp
@@ -27,7 +27,7 @@ namespace hardware {
 namespace nfc {
 
 void NfcExtns::getConfig(NfcConfig& config) {
-  uint64_t num = 0;
+  uint32_t num = 0;
   std::array<uint8_t, NXP_MAX_CONFIG_STRING_LEN> buffer;
   buffer.fill(0);
   int64_t retlen = 0;
